# api/swagger/soccer-manager.swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    properties:
      data: {}
    type: object
...
  CreateTeamTranslationRequest:
    properties:
      label:
        type: string
      locale:
        type: string
    required:
    - label
    - locale
    type: object
...
  DeleteTeamTranslationRequest:
    properties:
      locale:
        type: string
    required:
    - locale
    type: object
...
  TeamResponse:
    properties:
      budget:
        type: integer
      country_code:
        type: string
      id:
        type: integer
      name:
        type: string
      total_players:
        type: integer
      user_id:
        type: integer
    type: object
  TeamTranslation:
    additionalProperties:
      type: string
    type: object
...
  UpdateTeamRequest:
    properties:
      country_code:
        type: string
      name:
        type: string
    required:
    - country_code
    - name
    type: object
  UpdateTeamTranslationRequest:
    properties:
      label:
        type: string
      locale:
        type: string
    required:
    - label
    - locale
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
host: localhost:80
info:
  contact: {}
  description: Soccer manager api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Soccer Manager
  version: 1.0.0-alpha0
paths:
...
  /v1/teams:
    get:
      description: Returns a paginated list of teams
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/TeamResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: List teams
      tags:
      - team
  /v1/teams/{team_id}:
    get:
      description: Returns a single team's data by ID
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/TeamResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get team by ID
      tags:
      - team
  /v1/teams/{team_id}/locales:
    get:
      description: Returns a list of available translation locales
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/TeamResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get available locales for team by ID
      tags:
      - team
...
  /v1/teams/me:
    put:
      consumes:
      - application/json
      description: Updates the name and country code of the authenticated user's team
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Update team
      tags:
      - team
  /v1/teams/me/translations:
    delete:
      description: Deletes a team translation for the authenticated user's team
      parameters:
      - description: Delete details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DeleteTeamTranslationRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Delete team translation
      tags:
      - team
    get:
      description: Returns translations for the authenticated user's team
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/TeamTranslation'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: List translations
      tags:
      - team
    post:
      consumes:
      - application/json
      description: Creates a new translation for the authenticated user's team
      parameters:
      - description: Translation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateTeamTranslationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Create team translation
      tags:
      - team
    put:
      consumes:
      - application/json
      description: Updates an existing translation for the authenticated user's team
      parameters:
      - description: Translation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateTeamTranslationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Update team translation
      tags:
      - team
...
  /v1/users/{user_id}/team:
    get:
      description: Returns a single team by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/TeamResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get team by user ID
      tags:
      - team
...
schemes:
- http
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"